refactor: 포스트 생성 페이지 페이지 단으로 정의
