--- src/components/CreatePost.tsx
@@ -1,96 +1,0 @@
-import { useForm } from 'react-hook-form';
-import { useState } from 'react';
-import PrimaryButton from './common/PrimaryButton';
-import {
-  FormControl,
-  FormLabel,
-  Input,
-  Text,
-  Stack,
-  Spacer,
-  Textarea,
-  Center,
-  FormErrorMessage,
-} from '@chakra-ui/react';
-import { ChevronLeftIcon } from '@chakra-ui/icons';
-import PrimaryHeader from './common/PrimaryHeader';
-import PrimaryLink from './common/PrimaryLink';
-import { ROUTES } from '../constants/routes';
-import { useQueryPost } from '../hooks/useQueryPost';
-import UploadImage from '../components/common/UploadImage';
-
-type PostData = {
-  title: string;
-  // 후기 json 추가 시 타입 지정 예정
-};
-
-const CreatePost = () => {
-  const {
-    register,
-    handleSubmit,
-    formState: { errors },
-  } = useForm<PostData>();
-  const [selectImage, setSelectImage] = useState<File | null>(null);
-  const { createPoster } = useQueryPost();
-
-  const handleChange = (file: File) => {
-    setSelectImage(file);
-  };
-
-  const onSubmit = (data: PostData) => {
-    const formData = new FormData();
-    if (data) {
-      formData.append('title', data.title);
-      formData.append('channelId', import.meta.env.VITE_MAIN_CHANNELID);
-      selectImage && formData.append('image', selectImage);
-      createPoster.mutate(formData);
-    }
-  };
-
-  return (
-    <>
-      <PrimaryHeader>
-        <PrimaryLink router={ROUTES.MAIN} width={5}>
-          <ChevronLeftIcon boxSize={10} />
-        </PrimaryLink>
-        <Spacer />
-        <Text as="b" fontSize="2xl">
-          후기작성
-        </Text>
-        <Spacer />
-      </PrimaryHeader>
-      <Stack as="form" onSubmit={handleSubmit(onSubmit)} spacing={4}>
-        <FormControl mt={5} isInvalid={!!errors.title}>
-          <FormLabel>제목</FormLabel>
-          <Input
-            placeholder="제목을 입력하세요."
-            {...register('title', {
-              required: 'title을 적어주세요.',
-            })}
-          />
-          {errors.title && (
-            <FormErrorMessage>{errors.title.message}</FormErrorMessage>
-          )}
-        </FormControl>
-        <FormControl>
-          <FormLabel>여행지</FormLabel>
-          <PrimaryButton w="180px">카카오맵 불러오기</PrimaryButton>
-        </FormControl>
-        <FormControl>
-          <FormLabel>후기</FormLabel>
-          <Textarea />
-        </FormControl>
-        <FormControl>
-          <UploadImage handleOnChange={handleChange} />
-        </FormControl>
-        <Center>
-          <PrimaryButton type="submit" w="200px">
-            글 올리기
-          </PrimaryButton>
-        </Center>
-      </Stack>
-    </>
-  );
-};
-
-export default CreatePost;
